move structs to types.ts
