+++ cocos/rendering/custom/index.ts
@@ -33,8 +33,6 @@ import { CustomPipelineBuilder } from './custom-pipeline';
 let _pipeline: WebPipeline | null = null;
 
 export * from './types';
-export { Descriptor, DescriptorTypeOrder, DescriptorBlockFlattened, DescriptorBlockIndex } from './layout-graph';
-export { CopyPair, MovePair } from './render-graph';
 export * from './pipeline';
 
 export function createCustomPipeline (): Pipeline {
