+++ native/cocos/renderer/pipeline/custom/LayoutGraphFwd.h
@@ -38,12 +38,6 @@ namespace cc {
 
 namespace render {
 
-enum class DescriptorTypeOrder;
-
-struct Descriptor;
-struct DescriptorBlock;
-struct DescriptorBlockFlattened;
-struct DescriptorBlockIndex;
 struct DescriptorDB;
 struct RenderStageTag;
 struct RenderPhaseTag;
