+++ native/cocos/renderer/pipeline/custom/RenderCommonTypes.cpp
@@ -82,6 +82,77 @@ ComputeView::ComputeView(ComputeView const& rhs, const allocator_type& alloc)
   clearColor(rhs.clearColor),
   clearValueType(rhs.clearValueType) {}
 
+CopyPair::CopyPair(const allocator_type& alloc) noexcept
+: source(alloc),
+  target(alloc) {}
+
+CopyPair::CopyPair(ccstd::pmr::string sourceIn, ccstd::pmr::string targetIn, uint32_t mipLevelsIn, uint32_t numSlicesIn, uint32_t sourceMostDetailedMipIn, uint32_t sourceFirstSliceIn, uint32_t sourcePlaneSliceIn, uint32_t targetMostDetailedMipIn, uint32_t targetFirstSliceIn, uint32_t targetPlaneSliceIn, const allocator_type& alloc) noexcept // NOLINT
+: source(std::move(sourceIn), alloc),
+  target(std::move(targetIn), alloc),
+  mipLevels(mipLevelsIn),
+  numSlices(numSlicesIn),
+  sourceMostDetailedMip(sourceMostDetailedMipIn),
+  sourceFirstSlice(sourceFirstSliceIn),
+  sourcePlaneSlice(sourcePlaneSliceIn),
+  targetMostDetailedMip(targetMostDetailedMipIn),
+  targetFirstSlice(targetFirstSliceIn),
+  targetPlaneSlice(targetPlaneSliceIn) {}
+
+CopyPair::CopyPair(CopyPair&& rhs, const allocator_type& alloc)
+: source(std::move(rhs.source), alloc),
+  target(std::move(rhs.target), alloc),
+  mipLevels(rhs.mipLevels),
+  numSlices(rhs.numSlices),
+  sourceMostDetailedMip(rhs.sourceMostDetailedMip),
+  sourceFirstSlice(rhs.sourceFirstSlice),
+  sourcePlaneSlice(rhs.sourcePlaneSlice),
+  targetMostDetailedMip(rhs.targetMostDetailedMip),
+  targetFirstSlice(rhs.targetFirstSlice),
+  targetPlaneSlice(rhs.targetPlaneSlice) {}
+
+CopyPair::CopyPair(CopyPair const& rhs, const allocator_type& alloc)
+: source(rhs.source, alloc),
+  target(rhs.target, alloc),
+  mipLevels(rhs.mipLevels),
+  numSlices(rhs.numSlices),
+  sourceMostDetailedMip(rhs.sourceMostDetailedMip),
+  sourceFirstSlice(rhs.sourceFirstSlice),
+  sourcePlaneSlice(rhs.sourcePlaneSlice),
+  targetMostDetailedMip(rhs.targetMostDetailedMip),
+  targetFirstSlice(rhs.targetFirstSlice),
+  targetPlaneSlice(rhs.targetPlaneSlice) {}
+
+MovePair::MovePair(const allocator_type& alloc) noexcept
+: source(alloc),
+  target(alloc) {}
+
+MovePair::MovePair(ccstd::pmr::string sourceIn, ccstd::pmr::string targetIn, uint32_t mipLevelsIn, uint32_t numSlicesIn, uint32_t targetMostDetailedMipIn, uint32_t targetFirstSliceIn, uint32_t targetPlaneSliceIn, const allocator_type& alloc) noexcept // NOLINT
+: source(std::move(sourceIn), alloc),
+  target(std::move(targetIn), alloc),
+  mipLevels(mipLevelsIn),
+  numSlices(numSlicesIn),
+  targetMostDetailedMip(targetMostDetailedMipIn),
+  targetFirstSlice(targetFirstSliceIn),
+  targetPlaneSlice(targetPlaneSliceIn) {}
+
+MovePair::MovePair(MovePair&& rhs, const allocator_type& alloc)
+: source(std::move(rhs.source), alloc),
+  target(std::move(rhs.target), alloc),
+  mipLevels(rhs.mipLevels),
+  numSlices(rhs.numSlices),
+  targetMostDetailedMip(rhs.targetMostDetailedMip),
+  targetFirstSlice(rhs.targetFirstSlice),
+  targetPlaneSlice(rhs.targetPlaneSlice) {}
+
+MovePair::MovePair(MovePair const& rhs, const allocator_type& alloc)
+: source(rhs.source, alloc),
+  target(rhs.target, alloc),
+  mipLevels(rhs.mipLevels),
+  numSlices(rhs.numSlices),
+  targetMostDetailedMip(rhs.targetMostDetailedMip),
+  targetFirstSlice(rhs.targetFirstSlice),
+  targetPlaneSlice(rhs.targetPlaneSlice) {}
+
 } // namespace render
 
 } // namespace cc
