+++ native/cocos/renderer/pipeline/custom/RenderGraphJsb.cpp
@@ -35,73 +35,6 @@
 #include "cocos/renderer/pipeline/custom/RenderGraphJsb.h"
 #include "cocos/renderer/pipeline/custom/RenderGraphTypes.h"
 
-bool nativevalue_to_se(const cc::render::CopyPair &from, se::Value &to, se::Object *ctx) { // NOLINT
-    se::HandleObject obj(se::Object::createPlainObject());
-    se::Value        tmp;
-
-    nativevalue_to_se(from.source, tmp, ctx);
-    obj->setProperty("source", tmp);
-
-    nativevalue_to_se(from.target, tmp, ctx);
-    obj->setProperty("target", tmp);
-
-    nativevalue_to_se(from.mipLevels, tmp, ctx);
-    obj->setProperty("mipLevels", tmp);
-
-    nativevalue_to_se(from.numSlices, tmp, ctx);
-    obj->setProperty("numSlices", tmp);
-
-    nativevalue_to_se(from.sourceMostDetailedMip, tmp, ctx);
-    obj->setProperty("sourceMostDetailedMip", tmp);
-
-    nativevalue_to_se(from.sourceFirstSlice, tmp, ctx);
-    obj->setProperty("sourceFirstSlice", tmp);
-
-    nativevalue_to_se(from.sourcePlaneSlice, tmp, ctx);
-    obj->setProperty("sourcePlaneSlice", tmp);
-
-    nativevalue_to_se(from.targetMostDetailedMip, tmp, ctx);
-    obj->setProperty("targetMostDetailedMip", tmp);
-
-    nativevalue_to_se(from.targetFirstSlice, tmp, ctx);
-    obj->setProperty("targetFirstSlice", tmp);
-
-    nativevalue_to_se(from.targetPlaneSlice, tmp, ctx);
-    obj->setProperty("targetPlaneSlice", tmp);
-
-    to.setObject(obj);
-    return true;
-}
-
-bool nativevalue_to_se(const cc::render::MovePair &from, se::Value &to, se::Object *ctx) { // NOLINT
-    se::HandleObject obj(se::Object::createPlainObject());
-    se::Value        tmp;
-
-    nativevalue_to_se(from.source, tmp, ctx);
-    obj->setProperty("source", tmp);
-
-    nativevalue_to_se(from.target, tmp, ctx);
-    obj->setProperty("target", tmp);
-
-    nativevalue_to_se(from.mipLevels, tmp, ctx);
-    obj->setProperty("mipLevels", tmp);
-
-    nativevalue_to_se(from.numSlices, tmp, ctx);
-    obj->setProperty("numSlices", tmp);
-
-    nativevalue_to_se(from.targetMostDetailedMip, tmp, ctx);
-    obj->setProperty("targetMostDetailedMip", tmp);
-
-    nativevalue_to_se(from.targetFirstSlice, tmp, ctx);
-    obj->setProperty("targetFirstSlice", tmp);
-
-    nativevalue_to_se(from.targetPlaneSlice, tmp, ctx);
-    obj->setProperty("targetPlaneSlice", tmp);
-
-    to.setObject(obj);
-    return true;
-}
-
 bool nativevalue_to_se(const cc::render::ClearView &from, se::Value &to, se::Object *ctx) { // NOLINT
     se::HandleObject obj(se::Object::createPlainObject());
     se::Value        tmp;
@@ -117,94 +50,6 @@ bool nativevalue_to_se(const cc::render::ClearView &from, se::Value &to, se::Obj
 
     to.setObject(obj);
     return true;
-}
-
-template <>
-bool sevalue_to_native<cc::render::CopyPair>(const se::Value &from, cc::render::CopyPair *to, se::Object *ctx) { // NOLINT
-    SE_PRECONDITION2(from.isObject(), false, " Convert parameter to CopyPair failed !");
-
-    auto *obj = const_cast<se::Object *>(from.toObject());
-    bool ok = true;
-    se::Value field;
-    obj->getProperty("source", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->source), ctx);
-    }
-    obj->getProperty("target", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->target), ctx);
-    }
-    obj->getProperty("mipLevels", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->mipLevels), ctx);
-    }
-    obj->getProperty("numSlices", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->numSlices), ctx);
-    }
-    obj->getProperty("sourceMostDetailedMip", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->sourceMostDetailedMip), ctx);
-    }
-    obj->getProperty("sourceFirstSlice", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->sourceFirstSlice), ctx);
-    }
-    obj->getProperty("sourcePlaneSlice", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->sourcePlaneSlice), ctx);
-    }
-    obj->getProperty("targetMostDetailedMip", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->targetMostDetailedMip), ctx);
-    }
-    obj->getProperty("targetFirstSlice", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->targetFirstSlice), ctx);
-    }
-    obj->getProperty("targetPlaneSlice", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->targetPlaneSlice), ctx);
-    }
-    return ok;
-}
-
-template <>
-bool sevalue_to_native<cc::render::MovePair>(const se::Value &from, cc::render::MovePair *to, se::Object *ctx) { // NOLINT
-    SE_PRECONDITION2(from.isObject(), false, " Convert parameter to MovePair failed !");
-
-    auto *obj = const_cast<se::Object *>(from.toObject());
-    bool ok = true;
-    se::Value field;
-    obj->getProperty("source", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->source), ctx);
-    }
-    obj->getProperty("target", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->target), ctx);
-    }
-    obj->getProperty("mipLevels", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->mipLevels), ctx);
-    }
-    obj->getProperty("numSlices", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->numSlices), ctx);
-    }
-    obj->getProperty("targetMostDetailedMip", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->targetMostDetailedMip), ctx);
-    }
-    obj->getProperty("targetFirstSlice", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->targetFirstSlice), ctx);
-    }
-    obj->getProperty("targetPlaneSlice", &field, true);
-    if(!field.isNullOrUndefined()) {
-        ok &= sevalue_to_native(field, &(to->targetPlaneSlice), ctx);
-    }
-    return ok;
 }
 
 template <>
