+++ native/cocos/renderer/pipeline/custom/RenderGraphNames.h
@@ -56,9 +56,7 @@ inline const char* getName(const RasterSubpass& /*v*/) noexcept { return "Raster
 inline const char* getName(const SubpassGraph& /*v*/) noexcept { return "SubpassGraph"; }
 inline const char* getName(const RasterPass& /*v*/) noexcept { return "RasterPass"; }
 inline const char* getName(const ComputePass& /*v*/) noexcept { return "ComputePass"; }
-inline const char* getName(const CopyPair& /*v*/) noexcept { return "CopyPair"; }
 inline const char* getName(const CopyPass& /*v*/) noexcept { return "CopyPass"; }
-inline const char* getName(const MovePair& /*v*/) noexcept { return "MovePair"; }
 inline const char* getName(const MovePass& /*v*/) noexcept { return "MovePass"; }
 inline const char* getName(const RaytracePass& /*v*/) noexcept { return "RaytracePass"; }
 inline const char* getName(const QueueTag& /*v*/) noexcept { return "Queue"; }
